+++ src/renderer/store/files.ts
@@ -1,19 +1,26 @@
-import { SavedFile } from "@@/model";
-import { defineStore } from "pinia";
-import { ref } from "vue";
+import { SavedFile } from '@@/model'
+import { defineStore } from 'pinia'
+import { ref } from 'vue'
 import { Draft, create } from 'mutative'
-import { createConfig } from "@renderer/utils/config";
+import { createConfig } from '@renderer/utils/config'
 import { klona } from 'klona'
-import { SaveLocation } from "@@/save-location";
+import { SaveLocationValidator } from '@@/save-location'
+import { object, string, optional, record, InferInput, parse, ValiError } from 'valibot'
 
 export interface File {
   data: SavedFile
 }
 
-export interface FileRepo {
-  version: string
-  data: Record<string, SaveLocation>
-}
+export const FileRepoValidator = object({
+  version: optional(string(), '1.0.0'),
+  data: optional(record(string(), SaveLocationValidator), {})
+})
+
+// export interface FileRepo {
+//   version: string
+//   data: Record<string, SaveLocation>
+// }
+export type FileRepo = InferInput<typeof FileRepoValidator>
 
 export const useFile = (name: string) => {
   // const file = ref<File>()
@@ -22,22 +29,23 @@ export const useFile = (name: string) => {
 
   return {
     save,
-    load,
+    load
   }
 }
 
 const defaultFileRepo: FileRepo = {
-  version: "1.0.0",
+  version: '1.0.0',
   data: {}
 }
 
 export const useFiles = defineStore('files', () => {
-  const files = ref<FileRepo>(defaultFileRepo);
+  const files = ref<FileRepo>(defaultFileRepo)
 
   const { load: loadConfig, save: saveConfig } = createConfig<FileRepo>('projects')
 
   const update = async (callback: (state: Draft<FileRepo>) => void) => {
     files.value = create(files.value, callback)
+    console.log('files.value', files.value)
     await saveConfig(klona(files.value))
   }
 
@@ -45,7 +53,15 @@ export const useFiles = defineStore('files', () => {
     const data = await loadConfig()
 
     if (data.type === 'success') {
-      files.value = data.result.result
+      try {
+        files.value = parse(FileRepoValidator, data.result.result)
+      } catch (e) {
+        if (e instanceof ValiError) {
+          console.log("error", e.issues)
+        }
+        console.error('error', e)
+        files.value = defaultFileRepo
+      }
     } else {
       files.value = defaultFileRepo
     }
@@ -60,7 +76,8 @@ export const useFiles = defineStore('files', () => {
   const loadFile = (name: string) => {
     const { load, save } = createConfig<Record<string, File>>(name)
     return {
-      load, save
+      load,
+      save
     }
   }
 
@@ -71,6 +88,6 @@ export const useFiles = defineStore('files', () => {
     load,
     loadFile,
     update,
-    remove,
+    remove
   }
 })