+++ src/renderer/utils/fmt.ts
@@ -1,5 +1,5 @@
 export const fmt = {
   param: (value: string, variant?: 'primary' | 'secondary' | undefined, ifEmpty: string = "") => {
-    return `<div class="param ${variant ? variant : ''}">${value ? value : ifEmpty}</div>`
+    return `<span class="param ${variant ? variant : ''}">${value ? value : ifEmpty}</span>`
   }
 }